# swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0",
      "title": "User API",
      "description": "API for managing users in the system."
    },
    "host": "localhost:3000",
    "basePath": "/api",
    "schemes": ["http"],
    "paths": {
      "/users": {
        "get": {
          "summary": "Get all users",
          "description": "Returns a list of all users.",
          "responses": {
            "200": {
              "description": "A list of users.",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/User"
                }
              }
            }
          }
        },
        "post": {
          "summary": "Create a new user",
          "description": "Adds a new user to the system.",
          "parameters": [
            {
              "in": "body",
              "name": "user",
              "description": "The user to create.",
              "required": true,
              "schema": {
                "$ref": "#/definitions/User"
              }
            }
          ],
          "responses": {
            "201": {
              "description": "User created successfully."
            }
          }
        }
      },
      "/users/{id}": {
        "get": {
          "summary": "Get a user by ID",
          "description": "Returns a single user.",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "A single user.",
              "schema": {
                "$ref": "#/definitions/User"
              }
            },
            "404": {
              "description": "User not found."
            }
          }
        },
        "put": {
          "summary": "Update a user",
          "description": "Updates an existing user by ID.",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "type": "string"
            },
            {
              "in": "body",
              "name": "user",
              "description": "The updated user object.",
              "required": true,
              "schema": {
                "$ref": "#/definitions/User"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "User updated successfully."
            },
            "404": {
              "description": "User not found."
            }
          }
        },
        "delete": {
          "summary": "Delete a user",
          "description": "Deletes a user by ID.",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "204": {
              "description": "User deleted successfully."
            },
            "404": {
              "description": "User not found."
            }
          }
        }
      }
    },
    "definitions": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        },
        "required": ["name", "email"]
      }
    }
  }
  